test(matchLabels): add test for bug reported in [1]
[1]:https://github.com/godaddy/kubernetes-client/issues/90
